allow conversion between compatible dyn pointers
This allows conversion from `*dyn mut A` to `*dyn A` and from
`*dyn (mut) A` to `*dyn (mut) B` if `A` requires implementation of `B`.
The generation of vtables has been modified to make this simpler:
Vtables are now generated as a flat array of generic function pointers,
which are then casted at the call site to the correct type. If this ends
up having an effect on optimization, we can generate vtables the old
way, but instead of dumping all functions in each vtable, vtables can
embed other vtables inside them.